+++ src/index.rs
@@ -406,12 +406,16 @@ impl Index {
       .unwrap_or(0)
   }
 
-  pub(crate) fn height(&self) -> Result<Option<Height>> {
-    self.begin_read()?.height()
-  }
-
   pub(crate) fn block_count(&self) -> Result<u64> {
     self.begin_read()?.block_count()
+  }
+
+  pub(crate) fn block_height(&self) -> Result<Option<Height>> {
+    self.begin_read()?.block_height()
+  }
+
+  pub(crate) fn block_hash(&self, height: Option<u64>) -> Result<Option<BlockHash>> {
+    self.begin_read()?.block_hash(height)
   }
 
   pub(crate) fn blocks(&self, take: usize) -> Result<Vec<(u64, BlockHash)>> {
@@ -661,7 +665,7 @@ impl Index {
     }
   }
 
-  pub(crate) fn blocktime(&self, height: Height) -> Result<Blocktime> {
+  pub(crate) fn block_time(&self, height: Height) -> Result<Blocktime> {
     let height = height.n();
 
     match self.get_block_by_height(height)? {
@@ -838,35 +842,44 @@ impl Index {
       inscription_id,
     );
 
-    if let Some(sat) = sat {
-      if self.has_sat_index().unwrap() {
-        assert_eq!(
-          InscriptionId::load(
-            *rtx
-              .open_table(SAT_TO_INSCRIPTION_ID)
-              .unwrap()
-              .get(&sat)
-              .unwrap()
-              .unwrap()
-              .value()
-          ),
-          inscription_id,
-        );
-
-        // we do not track common sats or anything in the unbound output
-        if !Sat(sat).is_common() && satpoint.outpoint != unbound_outpoint() {
+    match sat {
+      Some(sat) => {
+        if self.has_sat_index().unwrap() {
+          // unbound inscriptions should not be assigned to a sat
+          assert!(satpoint.outpoint != unbound_outpoint());
           assert_eq!(
-            SatPoint::load(
+            InscriptionId::load(
               *rtx
-                .open_table(SAT_TO_SATPOINT)
+                .open_table(SAT_TO_INSCRIPTION_ID)
                 .unwrap()
                 .get(&sat)
                 .unwrap()
                 .unwrap()
                 .value()
             ),
-            satpoint,
+            inscription_id,
           );
+
+          // we do not track common sats (only the sat ranges)
+          if !Sat(sat).is_common() {
+            assert_eq!(
+              SatPoint::load(
+                *rtx
+                  .open_table(SAT_TO_SATPOINT)
+                  .unwrap()
+                  .get(&sat)
+                  .unwrap()
+                  .unwrap()
+                  .value()
+              ),
+              satpoint,
+            );
+          }
+        }
+      }
+      None => {
+        if self.has_sat_index().unwrap() {
+          assert!(satpoint.outpoint == unbound_outpoint())
         }
       }
     }
@@ -1002,21 +1015,21 @@ mod tests {
     {
       let context = Context::builder().args(["--height-limit", "0"]).build();
       context.mine_blocks(1);
-      assert_eq!(context.index.height().unwrap(), None);
+      assert_eq!(context.index.block_height().unwrap(), None);
       assert_eq!(context.index.block_count().unwrap(), 0);
     }
 
     {
       let context = Context::builder().args(["--height-limit", "1"]).build();
       context.mine_blocks(1);
-      assert_eq!(context.index.height().unwrap(), Some(Height(0)));
+      assert_eq!(context.index.block_height().unwrap(), Some(Height(0)));
       assert_eq!(context.index.block_count().unwrap(), 1);
     }
 
     {
       let context = Context::builder().args(["--height-limit", "2"]).build();
       context.mine_blocks(2);
-      assert_eq!(context.index.height().unwrap(), Some(Height(1)));
+      assert_eq!(context.index.block_height().unwrap(), Some(Height(1)));
       assert_eq!(context.index.block_count().unwrap(), 2);
     }
   }
@@ -2178,7 +2191,7 @@ mod tests {
           outpoint: unbound_outpoint(),
           offset: 0,
         },
-        Some(50 * COIN_VALUE),
+        None, // should not be on a sat
       );
 
       assert!(context
@@ -2355,7 +2368,7 @@ mod tests {
           outpoint: OutPoint { txid, vout: 0 },
           offset: 0,
         },
-        None,
+        Some(50 * COIN_VALUE),
       );
 
       context.index.assert_inscription_location(
@@ -2364,7 +2377,7 @@ mod tests {
           outpoint: OutPoint { txid, vout: 0 },
           offset: 50 * COIN_VALUE,
         },
-        None,
+        Some(100 * COIN_VALUE),
       );
 
       context.index.assert_inscription_location(
@@ -2373,7 +2386,7 @@ mod tests {
           outpoint: OutPoint { txid, vout: 0 },
           offset: 100 * COIN_VALUE,
         },
-        None,
+        Some(150 * COIN_VALUE),
       );
 
       assert_eq!(
@@ -2460,7 +2473,7 @@ mod tests {
           outpoint: OutPoint { txid, vout: 0 },
           offset: 0,
         },
-        None,
+        Some(50 * COIN_VALUE),
       );
 
       context.index.assert_inscription_location(
@@ -2567,7 +2580,7 @@ mod tests {
           outpoint: OutPoint { txid, vout: 0 },
           offset: 0,
         },
-        None,
+        Some(50 * COIN_VALUE),
       );
 
       context.index.assert_inscription_location(
@@ -2576,7 +2589,7 @@ mod tests {
           outpoint: OutPoint { txid, vout: 0 },
           offset: 50 * COIN_VALUE,
         },
-        None,
+        Some(100 * COIN_VALUE),
       );
 
       context.index.assert_inscription_location(
@@ -2719,7 +2732,7 @@ mod tests {
           },
           offset: 0,
         },
-        None,
+        Some(100 * COIN_VALUE),
       );
 
       assert_eq!(