+++ src/index/rtx.rs
@@ -3,7 +3,7 @@ use super::*;
 pub(crate) struct Rtx<'a>(pub(crate) redb::ReadTransaction<'a>);
 
 impl Rtx<'_> {
-  pub(crate) fn height(&self) -> Result<Option<Height>> {
+  pub(crate) fn block_height(&self) -> Result<Option<Height>> {
     Ok(
       self
         .0
@@ -26,5 +26,26 @@ impl Rtx<'_> {
         .map(|(height, _hash)| height.value() + 1)
         .unwrap_or(0),
     )
+  }
+
+  pub(crate) fn block_hash(&self, height: Option<u64>) -> Result<Option<BlockHash>> {
+    match height {
+      Some(height) => Ok(
+        self
+          .0
+          .open_table(HEIGHT_TO_BLOCK_HASH)?
+          .get(height)?
+          .map(|hash| BlockHash::load(*hash.value())),
+      ),
+      None => Ok(
+        self
+          .0
+          .open_table(HEIGHT_TO_BLOCK_HASH)?
+          .range(0..)?
+          .rev()
+          .next()
+          .map(|(_height, hash)| BlockHash::load(*hash.value())),
+      ),
+    }
   }
 }